✨ feat: 식당의 웨이팅 전체조회 ,특정 웨이팅 받기, 웨이팅 취소, 웨이팅 정보 수정, 단일 웨이팅 정보 조회 - 식당의 웨이팅 전체조회 : 식당의 웨이팅 리스트를 페이지네이션 해서 받아오는 기능 - 특정 웨이팅 받기 : 특정 웨이팅 번호를 받고 해당 웨이팅을 식당에 출입한 것으로 처리하는 기능(이후 번호 앞당겨짐) - 웨이팅 취소 : 특정 웨이팅 번호를 받고 해당 웨이팅을 리스트에서 삭제하는 기능(이후 번호 앞당겨짐) - 웨이팅 정보 수정 : 웨이팅의 '웨이팅 인원'과 '요청사항'을 수정 가능함 - 단일 웨이팅 정보 조회 : 단일 웨이팅의 정보를 조회 할 수 있는 기능

+++ src/main/java/com/sparta/goodbite/domain/waiting/service/WaitingService.java
@@ -1,18 +1,28 @@
 package com.sparta.goodbite.domain.waiting.service;
 
+import com.sparta.goodbite.domain.Customer.entity.Customer;
+import com.sparta.goodbite.domain.Customer.repository.CustomerRepository;
 import com.sparta.goodbite.domain.restaurant.entity.Restaurant;
 import com.sparta.goodbite.domain.restaurant.repository.RestaurantRepository;
-import com.sparta.goodbite.domain.user.entity.Customer;
-import com.sparta.goodbite.domain.user.repository.CustomerRepository;
 import com.sparta.goodbite.domain.waiting.dto.PostWaitingRequestDto;
 import com.sparta.goodbite.domain.waiting.dto.UpdateWaitingRequestDto;
 import com.sparta.goodbite.domain.waiting.dto.WaitingResponseDto;
 import com.sparta.goodbite.domain.waiting.entity.Waiting;
 import com.sparta.goodbite.domain.waiting.entity.Waiting.WaitingStatus;
-import com.sparta.goodbite.domain.waiting.entity.Waiting.WaitingType;
 import com.sparta.goodbite.domain.waiting.repository.WaitingRepository;
+import com.sparta.goodbite.exception.waiting.WaitingErrorCode;
+import com.sparta.goodbite.exception.waiting.WaitingException;
+import com.sparta.goodbite.exception.waiting.detail.WaitingNotFoundException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -26,21 +36,26 @@ public class WaitingService {
     public WaitingResponseDto createWaiting(
 //        UserDetailsImpl userDetails,
         PostWaitingRequestDto postWaitingRequestDto) {
-        Long waitingOrder = 1L; // 추후 List에서 마지막 번호 반환 후 +1해서 저장하도록 수정 필요
 
         Restaurant restaurant = restaurantRepository.findById(
             postWaitingRequestDto.getRestaurantId()).orElseThrow();
 
-        Customer customer = customerRepository.findById(1L)
+        System.out.println(restaurant.getId());
+
+        Customer customer = customerRepository.findById(3L)
             .orElseThrow(); // 추후 UserDetails 받고 수정 필요
+
+        Long LastOrderNumber = findLastOrderNumber(restaurant);
+
+        System.out.println(LastOrderNumber);
 
         Waiting waiting = new Waiting(
             restaurant,
             customer,
-            waitingOrder,
+            LastOrderNumber + 1,
             WaitingStatus.WAITING, //생성 시 무조껀 Waiting
             postWaitingRequestDto.getPartySize(),
-            WaitingType.OFFLINE,
+            postWaitingRequestDto.getWaitingType(),
             postWaitingRequestDto.getDemand());
         System.out.println(waiting.getWaitingType());
 
@@ -49,28 +64,59 @@ public class WaitingService {
         return WaitingResponseDto.from(waiting, restaurant.getName());
     }
 
-//    public WaitingResponseDto getWaiting(Long waitingId) {
-//
-//        return WaitingResponseDto.from(waitingRepository.findById(waitingId).orElseThrow());
-//    }
+    // 단일 조회용 메서드
+    public WaitingResponseDto getWaiting(Long waitingId) {
 
-    //가게 주인용 api
-    // 해당 메서드 동작 시, 가게의 id가 들어간 orders가 하나씩 줄게 된다.
-    public WaitingResponseDto updateWaitingOrders(
-//        UserDetailsImpl userDetails,
-        Long waitingId,
-        UpdateWaitingRequestDto updateWaitingRequestDto) {
-
+        Waiting waiting = waitingRepository.findById(waitingId)
+            .orElseThrow(() -> new WaitingNotFoundException(
+                WaitingErrorCode.WAITING_NOT_FOUND));
+        return WaitingResponseDto.from(waiting, waiting.getRestaurant().getName());
     }
 
+    //    //가게 주인용 api
+//    //해당 메서드 동작 시, 가게의 id가 들어간 orders가 하나씩 줄게 된다.
+    // restaurant id에 맞는 Waiting들의 order를 하나씩 줄인다.
+    @Transactional
+    public void reduceAllWaitingOrders(
+//        UserDetailsImpl userDetails,
+        Long restaurantId) {
+        List<Waiting> waitingList = waitingRepository.findByRestaurantId(restaurantId);
+        if (waitingList.isEmpty()) {
+            throw new WaitingException(WaitingErrorCode.WAITING_NOT_FOUND);
+        }
 
-    // 손님용 api
+        List<Waiting> waitingArrayList = new ArrayList<>();
+
+        for (Waiting waiting : waitingList) {
+            waiting.reduceWaitingOrder();
+            waitingArrayList.add(waiting);
+            if (waiting.getWaitingOrder() == 0) {
+                // 그리고 여기서 알람이나 그런거 해야 함
+                System.out.println("0번은 입장하세요~");
+                //알람!
+                deleteWaiting(waiting.getId());
+            }
+        }
+        // 쿼리가 계속 나감...
+        // 한꺼번에 범위로 줄일 수 있음
+        waitingRepository.saveAll(waitingArrayList);
+    }
+
+    // 웨이팅 하나만 삭제하고 뒤 웨이팅 숫자 하나씩 감소
+    @Transactional
+    public void reduceOneWaitingOrders(Long waitingId) {
+        reduceWaitingOrders(waitingId, "reduce");
+    }
+
+    // 가게용 api
     // 예약 인원수와 요청사항만 변경 가능함 ( 추후 합의를 통해 ?건 이하의 순서일 때는 수정하지 못하도록 로직 수정 필요)
     public WaitingResponseDto updateWaiting(Long waitingId,
 //        UserDetailsImpl userDetails,
         UpdateWaitingRequestDto updateWaitingRequestDto) {
 
-        Waiting waiting = waitingRepository.findById(waitingId).orElseThrow();
+        Waiting waiting = waitingRepository.findById(waitingId)
+            .orElseThrow(() -> new WaitingNotFoundException(
+                WaitingErrorCode.WAITING_NOT_FOUND));
         String restaurantName = waiting.getRestaurant().getName();
 
         waiting.update(updateWaitingRequestDto.getPartySize(), updateWaitingRequestDto.getDemand());
@@ -79,17 +125,76 @@ public class WaitingService {
         return WaitingResponseDto.from(waiting, restaurantName);
     }
 
-
     // 삭제 시, 메서드를 호출한 유저의 id와 취소하고자 하는 가게 id를 받아야 하는거 아닌가?
     // 프론트에서 어떤값을 주느냐에 따라 달라질 것 같다.
     public WaitingResponseDto deleteWaiting(Long waitingId) {
 //        UserDetailsImpl userDetails,
-        Waiting waiting = waitingRepository.findById(waitingId).orElseThrow();
+        Waiting waiting = waitingRepository.findById(waitingId)
+            .orElseThrow(() -> new WaitingNotFoundException(
+                WaitingErrorCode.WAITING_NOT_FOUND));
+
         String restaurantName = waiting.getRestaurant().getName();
 
-        waitingRepository.delete(waiting);
+        reduceWaitingOrders(waitingId, "delete");
+
         return WaitingResponseDto.from(waiting, restaurantName);
     }
 
+    public Long findLastOrderNumber(
+        Restaurant restaurant) {
+        if (!waitingRepository.findByRestaurantId(restaurant.getId()).isEmpty()) {
+            // 해당하는 레스토랑에 예약이 하나라도 존재한다면
+            return waitingRepository.findMaxWaitingOrderByRestaurantId(
+                restaurant.getId());
+        } else {
+            // 해당하는 레스토랑에 예약이 하나도 없다
+            return 0L;
+        }
+    }
 
+    public Page<WaitingResponseDto> getWaitingsByRestaurantId(Long restaurantId,
+        Pageable pageable) {
+        Page<Waiting> waitingPage = waitingRepository.findByRestaurantId(restaurantId, pageable);
+        List<WaitingResponseDto> waitingResponseDtos = waitingPage.stream()
+            .map(this::convertToDto)
+            .collect(Collectors.toList());
+        return new PageImpl<>(waitingResponseDtos, pageable, waitingPage.getTotalElements());
+    }
+
+    private WaitingResponseDto convertToDto(Waiting waiting) {
+        return WaitingResponseDto.from(waiting, waiting.getRestaurant().getName());
+    }
+
+    private void reduceWaitingOrders(Long waitingId, String type) {
+        Waiting waitingOne = waitingRepository.findById(waitingId)
+            .orElseThrow(() -> new WaitingException(WaitingErrorCode.WAITING_NOT_FOUND));
+        List<Waiting> waitingList = waitingRepository.findByRestaurantId(
+            waitingOne.getRestaurant().getId());
+
+        boolean flag = false;
+        List<Waiting> waitingArrayList = new ArrayList<>();
+
+        for (Waiting waiting : waitingList) {
+            if (Objects.equals(waiting.getId(), waitingId)) {
+                //여기서 알람 메서드
+                if (type.equals("delete")) {
+                    System.out.println("삭제 성공 알림");
+                } else if (type.equals("reduce")) {
+                    System.out.println("식당 웨이팅 받기 성공 알림");
+                }
+                System.out.println("0번은 입장하세요!");
+                waitingRepository.delete(waiting);
+                flag = true;
+            } else if (flag) {
+                waiting.reduceWaitingOrder();
+                waitingArrayList.add(waiting);
+            }
+        }
+        if (!flag) {
+            throw new WaitingNotFoundException(
+                WaitingErrorCode.WAITING_NOT_FOUND);
+        }
+        // 쿼리 최적화
+        waitingRepository.saveAll(waitingArrayList);
+    }
 }
